增加数据库迁移回滚锁开关，防止误删除

+++ database/migrations/2020_12_05_002103_sys_role.php
@@ -37,6 +37,8 @@ class SysRole extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists($this->tablename);
+        if (true === COMM_DBLOCK) {
+            Schema::dropIfExists($this->tablename);
+        }
     }
 }
